requirements.txt
fastapi
uvicorn
transformers
torch
torchvision
pillow
python-multipart
